// yc-prep-frontend-main/src/utils/questions/types.ts
export type QuestionCategory =
  | "Startup 101"
  | "Product Development"
  | "Management Operations"
  | "Growth Strategies"
  | "Market and Customers"
  | "Founder Traits"
  | "Industry knowledge";

export type QuestionOptionState =
  | "non-selected"
  | "selected"
  | "correct"
  | "incorrect";